feat: Asignaturas test

+++ routes/web.php
@@ -16,8 +16,8 @@ use App\Http\Middleware\CheckStudentRole;
 
     Route::get('/', [DestinosController::class, "index"])->name('destinos.admin')->middleware(CheckAdminRole::class);
     Route::get('/show/{id}', [DestinosController::class, "show"]);
-    Route::get('/create', [DestinosController::class, "create"]);
-    Route::post('/store', [DestinosController::class, "store"]);
+    Route::get('/create', [DestinosController::class, "create"])->middleware(CheckAdminRole::class);;
+    Route::post('/store', [DestinosController::class, "store"])->middleware(CheckAdminRole::class);;
     Route::get('/edit/{id}', [DestinosController::class, "edit"]);
     Route::get('/delete/{id}', [DestinosController::class, "delete"]);
     Route::get('/estudiante', [EstudianteController::class, "index"])->name('estudiante')->middleware(CheckStudentRole::class);
@@ -25,26 +25,26 @@ use App\Http\Middleware\CheckStudentRole;
 
     //Route::resource('asignaturas', AsignaturaController::class);
 
-    Route::get('/asignaturas', [AsignaturaController::class, "index"])->name('asignaturas.index');
-    Route::get('/asignaturas/create', [AsignaturaController::class, "create"])->name('asignaturas.create');
-    Route::post('/asignaturas/store', [AsignaturaController::class, "store"])->name('asignaturas.store');
-    Route::get('/asignaturas/{id}', [AsignaturaController::class, "searchByCity"])->name('asignatura');
-    Route::get('/asignaturas/{id}/edit', [AsignaturaController::class, "edit"])->name('asignaturas.edit');
-    Route::patch('/asignaturas/{id}', [AsignaturaController::class, "update"])->name('asignaturas.update');
-    Route::delete('/asignaturas/{id}', [AsignaturaController::class, "destroy"])->name('asignaturas.destroy');
+    Route::get('/asignaturas', [AsignaturaController::class, "index"])->name('asignaturas.index')->middleware(CheckAdminRole::class);;
+    Route::get('/asignaturas/create', [AsignaturaController::class, "create"])->name('asignaturas.create')->middleware(CheckAdminRole::class);
+    Route::post('/asignaturas/store', [AsignaturaController::class, "store"])->name('asignaturas.store')->middleware(CheckAdminRole::class);
+    Route::get('/asignaturas/{id}', [AsignaturaController::class, "searchByCity"])->name('asignatura')->middleware(CheckAdminRole::class);
+    Route::get('/asignaturas/{id}/edit', [AsignaturaController::class, "edit"])->name('asignaturas.edit')->middleware(CheckAdminRole::class);
+    Route::patch('/asignaturas/{id}', [AsignaturaController::class, "update"])->name('asignaturas.update')->middleware(CheckAdminRole::class);
+    Route::delete('/asignaturas/{id}', [AsignaturaController::class, "destroy"])->name('asignaturas.destroy')->middleware(CheckAdminRole::class);
     //Route::get('/asignaturas/create', [AsignaturaController::class, "create"])->name('asignaturas.create');
 
-    Route::get('/solicitudes', [SolicitudController::class, "index"])->name('solicitud');
-    Route::get('/solicitudesAprobadas', [SolicitudController::class, "solicitudesAprobadas"])->name('solicitudAprobadas');
+    Route::get('/solicitudes', [SolicitudController::class, "index"])->name('solicitud')->middleware(CheckAdminRole::class);
+    Route::get('/solicitudesAprobadas', [SolicitudController::class, "solicitudesAprobadas"])->name('solicitudAprobadas')->middleware(CheckAdminRole::class);
     Route::get('/solicitud/{idDestino}', [SolicitudController::class, "store"])->name('solicitud.store');
     Route::post('/cancelarSolicitud/{destinoId}', [SolicitudController::class, 'cancelarSolicitud'])->name('solicitudes.cancelar');
     Route::post('/solicitudes/renovar/{destinoId}', [SolicitudController::class, 'renovarSolicitud'])->name('solicitudes.renovar');
-    Route::post('/solicitudes/aprobar/{id}', [SolicitudController::class, 'aprobar'])->name('solicitudes.aprobar');
+    Route::post('/solicitudes/aprobar/{id}', [SolicitudController::class, 'aprobar'])->name('solicitudes.aprobar')->middleware(CheckAdminRole::class);
 
-    Route::get('/createDest',[DestinosController::class, "create"]);
-    Route::post('/storeDest',[DestinosController::class, "store"]);
-    Route::get('/ModDest/{id}',[DestinosController::class, "edit"]);
-    Route::put('/ModDest/{id}',[DestinosController::class, "update"]);
+    Route::get('/createDest',[DestinosController::class, "create"])->middleware(CheckAdminRole::class);
+    Route::post('/storeDest',[DestinosController::class, "store"])->middleware(CheckAdminRole::class);
+    Route::get('/ModDest/{id}',[DestinosController::class, "edit"])->middleware(CheckAdminRole::class);
+    Route::put('/ModDest/{id}',[DestinosController::class, "update"])->middleware(CheckAdminRole::class);
